Nature made the natural numbers; All else is the work of women

+++ 0x02-functions_nested_loops/101-natural.c
@@ -8,7 +8,7 @@
 
 int main(void)
 {
-	int i, sum;
+	unsigned long int i, sum;
 
 	sum = 0;
